packages/contracts/requirements.txt
eth-brownie>=1.10.0
